xstatic-package for Google's Material Design Icons
MDI provides Google's Material Design Icons in SCSS format
We plan on using this in Horizon to provide a theme that
implements Google's Material Design.

The 'material' theme gives an example of how to make use of a 3rd party theme
using the theming functionality.  In addition to incorporating a theme,
loaded as a static asset using requirements.txt, it also gives examples of how
to cleanly override styles, variables, icon fonts and Django templates.

This theme is replacing 'blue' as the example of how to use a theme other than
'default'.

Partially-Implements: blueprint horizon-theme-css-reorg
Change-Id: I356ce30e64daa13dd9ae8e5ed940b6b15982bd66