Fix description metadata
There doesn't seem to be a "summary" option, but "description" and
"long_description" are different. Update the metadata to avoid warnings
about the unknown "summary" and so that when the release announce script
asks for the package description it gets the one line version instead of
the longer one.

Had to bump the build version since there's already a tag for 1.6.50.1

Change-Id: I242418b4a94e57d926bafa0d537d3558d861c686